Refactor shared
