+++ src/features/popover/components/popoverNew/popoverNew.tsx
@@ -1,7 +1,7 @@
 import { type FC, useState, type ReactNode } from 'react';
 import { usePopper } from 'react-popper';
 import { useClickOutside } from '@shared/lib/hooks/useClickOutside';
-import { Portal } from '@shared/ui/portal/portal';
+import { Portal } from '@shared/ui/portal/portal/portal';
 
 interface PopoverProps {
     isOpen: boolean
