+++ src/shared/layouts/appBody/appBody.tsx
@@ -5,12 +5,10 @@ interface AppBodyProps {
     children: ReactNode
 }
 
-const AppBody = ({ children }: AppBodyProps) => {
+export const AppBody = ({ children }: AppBodyProps) => {
     return (
         <div className="app__body">
             <div className="content__container">{children}</div>
         </div>
     );
 };
-
-export default AppBody;
