+++ src/shared/layouts/index.ts
@@ -1,5 +1,3 @@
-import AppBody from './appBody/appBody';
-import AppHeader from './appHeader/appHeader';
-import AppLayout from './appLayout/appLayout';
-
-export { AppBody, AppHeader, AppLayout };
+export { AppBody } from './appBody/appBody';
+export { AppHeader } from './appHeader/appHeader';
+export { AppLayout } from './appLayout/appLayout';
